+++ src/index.ts
@@ -6,6 +6,7 @@ export { MemoryStorage } from './storages/MemoryStorage'
 export { RedisStorage } from './storages/RedisStorage'
 export type { RedisLikeClient } from './storages/RedisStorage'
 export { murmurhash3_32 } from './murmurhash'
+export { estimateSimilarity } from './Hash'
 export type { DocumentId, LshOptions, Query, QueryResult, ShingleUnit, StorageAdapter } from './types'
 
 /** Convenience factory, equivalent to `new Lsh(options)`. */